moved utils -> javascriptHelpers, moved commandValidInChannel to discordHelpers module
