+++ utils/javascriptHelpers.js
